Create plugin isolated traits
